refactor: tab carousel, 색상칩 분리 및 product-detail 적용
