+++ src/components/navigation/collectionPage.js
@@ -1,5 +1,6 @@
 import { collections } from "../../data/collection.js";
 import { renderNavigation } from "./utils/renderNavigation.js";
+import { initTabCarousel } from "../carousel/tabCarousel.js";
 
 const navigationList = document.getElementById("navigationList");
 
@@ -8,3 +9,6 @@ renderNavigation(navigationList, collections, {
   getLabel: (item) => item.name,
   getSlug: (item) => item.slug,
 });
+
+const tabsWrapper = document.getElementById("tabs-wrapper");
+initTabCarousel(tabsWrapper);
