+++ src/data/colors.js
@@ -1,0 +1,23 @@
+export const productColors = {
+  "0PEEZ09X6ASH7": {
+    activeIndex: 0,
+    chips: [
+      {
+        label: "레드 / 레드",
+        background:
+          "linear-gradient(135deg, rgb(194, 10, 64) 50%, rgb(194, 10, 64) 50%)",
+      },
+      {
+        label: "블루 / 블랙",
+        background:
+          "linear-gradient(135deg, rgb(31, 70, 164) 50%, rgb(17, 17, 17) 50%)",
+      },
+      {
+        label: "그레이 / 블랙",
+        background:
+          "linear-gradient(135deg, rgb(136, 129, 121) 50%, rgb(17, 17, 17) 50%)",
+      },
+    ],
+  },
+  // 추가 상품은 API 연동 시 채울 예정
+};
